Conflicts:
	gui/GameArena.cpp

+++ network/client/MapClient.cpp
@@ -90,7 +90,7 @@ void MapClient::updateMap(QByteArray& updateBlock) {
     quint8 nbRemovedBonus;
     updateIn >> nbRemovedBonus;
     for(quint8 i = 0; i < nbRemovedBonus; i++) {
-        qint16 x, y;
+        qint8 x, y;
         updateIn >> x >> y;
         emit sigRemoveBonus(x, y);
     }
